drm: per device prime dma buf hash
Dave's original prime patches had a per file lookup (currently a list,
but should probably be a hash as mentioned in the comment). This is
handy for when an importer tries to import the same BO multiple times as
it will prevent object duplication in that case. The problem is it does
not account for multiple importers. If an exporter passes a dma-buf fd
to multiple importers, a new object would be created for each one, which
is bad.

To handle multiple importers, we create a per device dma buf hash table
where the key is the dma_buf pointer, and the value is the gem object.
There is also a lock introduced to help with the various ref count
dances (and subsequent hash list traversals).

Cc: Daniel Vetter <daniel.vetter@ffwll.ch>
Cc: Dave Airlie <airlied@redhat.com>
Signed-off-by: Ben Widawsky <ben@bwidawsk.net>